+++ test/ModelMixins/TableMixinSpec.ts
@@ -1043,6 +1043,8 @@ describe("TableMixin", function () {
         }
       ];
 
+      console.log(item);
+
       styles.forEach((style, index) => {
         const feature = mapItem.entities.values[index];
         expect(