+++ test/Models/Catalog/esri/ArcGisFeatureServerCatalogItemSpec.ts
@@ -2,12 +2,15 @@ import i18next from "i18next";
 import { configure, runInAction } from "mobx";
 import Color from "terriajs-cesium/Source/Core/Color";
 import _loadWithXhr from "../../../../lib/Core/loadWithXhr";
+import ProtomapsImageryProvider from "../../../../lib/Map/ImageryProvider/ProtomapsImageryProvider";
+import { ProtomapsArcGisPbfSource } from "../../../../lib/Map/Vector/Protomaps/ProtomapsArcGisPbfSource";
 import { isDataSource } from "../../../../lib/ModelMixins/MappableMixin";
 import ArcGisFeatureServerCatalogItem, {
   convertEsriPointSizeToPixels
 } from "../../../../lib/Models/Catalog/Esri/ArcGisFeatureServerCatalogItem";
 import CommonStrata from "../../../../lib/Models/Definition/CommonStrata";
 import Terria from "../../../../lib/Models/Terria";
+import TableColumnType from "../../../../lib/Table/TableColumnType";
 
 configure({
   enforceActions: "observed",
@@ -30,6 +33,9 @@ describe("ArcGisFeatureServerCatalogItem", function () {
 
   const featureServerUrlStyleLines =
     "http://example.com/arcgis/rest/services/styles/FeatureServer/0";
+
+  const featureServerUrlTiled =
+    "http://example.com/arcgis/rest/services/tiled/FeatureServer/0";
 
   const featureServerUrlMulti =
     "http://example.com/arcgis/rest/services/Water_Network_Multi/FeatureServer/2";
@@ -69,6 +75,9 @@ describe("ArcGisFeatureServerCatalogItem", function () {
       } else if (originalUrl.match("styles/FeatureServer")) {
         url = url.replace(/FeatureServer\/0\/?\?.*/i, "lines.json");
         args[0] = "test/ArcGisFeatureServer/styles/" + url;
+      } else if (originalUrl.match("tiled/FeatureServer")) {
+        url = url.replace(/FeatureServer\/0\/?\?.*/i, "act-veg-map.json");
+        args[0] = "test/ArcGisFeatureServer/Tiled/" + url;
       } else if (originalUrl.match("Water_Network_Multi/FeatureServer")) {
         // We're getting this feature service in multiple requests, so we need to return different data on subsequent
         // calls
@@ -116,6 +125,11 @@ describe("ArcGisFeatureServerCatalogItem", function () {
       }
     });
 
+    it("defines min/max scale denominator", function () {
+      expect(item.minScaleDenominator).toEqual(0);
+      expect(item.maxScaleDenominator).toEqual(0);
+    });
+
     it("supports zooming to extent", function () {
       expect(item.disableZoomTo).toBeFalsy();
     });
@@ -139,11 +153,36 @@ describe("ArcGisFeatureServerCatalogItem", function () {
     });
   });
 
-  describe("loadMapItems", function () {
+  describe("after loading metadata - layer with min/max extent", function () {
+    beforeEach(async function () {
+      runInAction(() => {
+        item.setTrait("definition", "url", featureServerUrl2);
+      });
+      await item.loadMetadata();
+    });
+
+    it("defines a rectangle", function () {
+      expect(item.rectangle).toBeDefined();
+      if (item.rectangle) {
+        expect(item.rectangle.west).toEqual(150.01467801477867);
+        expect(item.rectangle.south).toEqual(-34.33741590494015);
+        expect(item.rectangle.east).toEqual(151.34389995591835);
+        expect(item.rectangle.north).toEqual(-32.99606922215397);
+      }
+    });
+
+    it("defines min/max scale denominator", function () {
+      expect(item.minScaleDenominator).toEqual(36111);
+      expect(item.maxScaleDenominator).toEqual(18489298);
+    });
+  });
+
+  describe("loadMapItems - geojson query", function () {
     it("properly loads a single layer", async function () {
       runInAction(() => {
         item.setTrait(CommonStrata.definition, "url", featureServerUrl);
         item.setTrait(CommonStrata.definition, "maxFeatures", 20);
+        item.setTrait(CommonStrata.definition, "tileRequests", false);
       });
 
       await item.loadMapItems();
@@ -165,6 +204,7 @@ describe("ArcGisFeatureServerCatalogItem", function () {
       runInAction(() => {
         item.setTrait(CommonStrata.definition, "url", featureServerUrlMulti);
         item.setTrait(CommonStrata.definition, "featuresPerRequest", 10);
+        item.setTrait(CommonStrata.definition, "tileRequests", false);
       });
 
       await item.loadMapItems();
@@ -179,6 +219,105 @@ describe("ArcGisFeatureServerCatalogItem", function () {
 
       // 1 call for metadata, and 2 calls for features
       expect(xhrSpy).toHaveBeenCalledTimes(3);
+    });
+  });
+
+  describe("loadMapItems - tiled query", function () {
+    it("properly loads a layer", async function () {
+      runInAction(() => {
+        item.setTrait(CommonStrata.definition, "url", featureServerUrlTiled);
+      });
+
+      await item.loadMapItems();
+
+      expect(item.tileRequests).toBeTruthy();
+      expect(item.maxTiledFeatures).toEqual(100000);
+      expect(item.featuresPerTileRequest).toEqual(4000);
+      expect(item.maxRecordCountFactor).toEqual(1);
+      expect(item.supportsQuantization).toEqual(true);
+      expect(item.objectIdField).toEqual("OBJECTID");
+      expect(item.outFields).toEqual(["OBJECTID", "class"]);
+
+      expect(item.mapItems.length).toEqual(1);
+      const mapItem = item.mapItems[0];
+
+      expect("imageryProvider" in mapItem).toBeTruthy();
+      if ("imageryProvider" in mapItem) {
+        const imageryProvider =
+          mapItem.imageryProvider as ProtomapsImageryProvider;
+
+        expect(
+          imageryProvider.source instanceof ProtomapsArcGisPbfSource
+        ).toBeTruthy();
+      }
+
+      expect(item.dataColumnMajor).toEqual([
+        ["OBJECTID"],
+        ["GlobalID"],
+        ["actGUID"],
+        ["vegCommunity"],
+        ["umcID"],
+        ["tecACT"],
+        ["tecEPBC"],
+        ["tecID"],
+        ["actConservationStatus"],
+        ["nationalConservationStatus"],
+        ["treesp1"],
+        ["treesp2"],
+        ["treesp3"],
+        ["shrubsp1"],
+        ["shrubsp2"],
+        ["shrubsp3"],
+        ["groundsp1"],
+        ["groundsp2"],
+        ["groundsp3"],
+        ["hectares"],
+        ["height_mean"],
+        ["height_stdev"],
+        ["canopyCover"],
+        ["underCover"],
+        ["structure"],
+        ["formation"],
+        ["class"],
+        ["grassyStructure"],
+        ["lastVegCommunity"],
+        ["landscape"],
+        ["lossReason"],
+        ["lossYear"],
+        ["Shape__Area"],
+        ["Shape__Length"]
+      ]);
+      expect(item.activeTableStyle.colorColumn?.name).toEqual("class");
+      expect(item.activeTableStyle.colorColumn?.type).toEqual(
+        TableColumnType.text
+      );
+      expect(item.activeTableStyle.tableColorMap.type).toEqual("enum");
+    });
+
+    it("ProtomapsImageryProvider - ArcGisPbfSource", async function () {
+      runInAction(() => {
+        item.setTrait(CommonStrata.definition, "url", featureServerUrlTiled);
+      });
+
+      await item.loadMapItems();
+
+      const mapItem = item.mapItems[0];
+
+      expect("imageryProvider" in mapItem).toBeTruthy();
+      if ("imageryProvider" in mapItem) {
+        const imageryProvider =
+          mapItem.imageryProvider as ProtomapsImageryProvider;
+
+        expect(
+          imageryProvider.source instanceof ProtomapsArcGisPbfSource
+        ).toBeTruthy();
+
+        // TODO
+        // Test get tile - with single request
+        // Test get tile - with multiple requests
+        // Test get tile - look at process features...
+        // Test pickFeautres
+      }
     });
   });
 
@@ -220,6 +359,7 @@ describe("ArcGisFeatureServerCatalogItem", function () {
           featureServerUrlStyleLines
         );
         item.setTrait(CommonStrata.definition, "maxFeatures", 20);
+        item.setTrait(CommonStrata.definition, "tileRequests", false);
       });
       await item.loadMapItems();
 
@@ -441,66 +581,91 @@ describe("ArcGisFeatureServerCatalogItem", function () {
       expect(tableStyle.outlineStyleMap.column?.name).toBe("id1");
 
       expect(tableStyle.outlineStyleMap.traitValues.null).toEqual({
+        color: "rgb(252,146,31)",
+        dash: [6, 6],
         width: 6
       });
 
       expect(tableStyle.outlineStyleMap.traitValues.enum).toEqual([
         {
+          color: "rgb(237,81,81)",
+          dash: [],
           legendTitle: "1",
           width: 2,
           value: "1"
         },
         {
+          color: "rgb(20,158,206)",
+          dash: [2, 4],
           legendTitle: "2",
           width: 2,
           value: "2"
         },
         {
+          color: "rgb(167,198,54)",
+          dash: [6, 6],
           legendTitle: "3",
           width: 2,
           value: "3"
         },
         {
+          color: "rgb(158,85,156)",
+          dash: [6, 3, 1, 3],
           legendTitle: "4",
           width: 2,
           value: "4"
         },
         {
+          color: "rgb(252,146,31)",
+          dash: [6, 3, 1, 3, 1, 3],
           legendTitle: "5",
           width: 2,
           value: "5"
         },
         {
+          color: "rgb(255,222,62)",
+          dash: [8, 4],
           legendTitle: "6",
           width: 2,
           value: "6"
         },
         {
+          color: "rgb(247,137,216)",
+          dash: [8, 3, 1, 3],
           legendTitle: "7",
           width: 2,
           value: "7"
         },
         {
+          color: "rgb(183,129,74)",
+          dash: [4, 4],
           legendTitle: "8",
           width: 2,
           value: "8"
         },
         {
+          color: "rgb(60,175,153)",
+          dash: [1, 2],
           legendTitle: "9",
           width: 2,
           value: "9"
         },
         {
+          color: "rgb(107,107,214)",
+          dash: [4, 2, 1, 2],
           legendTitle: "10",
           width: 2,
           value: "10"
         },
         {
+          color: "rgb(181,71,121)",
+          dash: [4, 2, 1, 2, 1, 2],
           legendTitle: "11",
           width: 2,
           value: "11"
         },
         {
+          dash: [],
           legendTitle: "12",
           width: 2,
           value: "12"